Adopt ~/sw/{sh.local,sh,bin} in place of ~/{sh,bin}
What I have been doing is making ~/sh a link to $DROPBOX/sh.pretty-name,
because most things that I put in ~/sh are system specific. Now ~/sw/sh
is a link to $DROPBOX/sh.common and ~/sw/sh.local is a link to
$DROPBOX/sh.pretty-name. Likewise, sh.local comes before sh in path.

sh comes before bin in path, so that scripts may override and/or wrap
binaries.