Remove .sh/services and support code.
This was kind of useful when I used to live the laptop life and want
different sessions based on work environment. Today, I tend to use
Android devices and fixed machines (e.g. work and home desktops).

Much of the functionality for this is no longer wanted and has not been
used in years.