Fix dump_path on FreeBSD.
Apparently FreeBSD sed doesn't allow using constructs like '\n' in the
substitution the way GNU sed does.

I've changed this to convert it to space and use a for loop to print it.
That has issues with PATH entries with spaces in them but that's likely
something unlikely off Windows or DOS, and I don't care about Windows or
DOS.