Add script to rotate screens.
  - rotate-screens: rotate first display.
  - rotate-screens --all: rotate all displays.
  - rotate-screens FOO BAR: rotate FOO and BAR.

If orientation is normal then they are rotated right. Might be nice to
have a flag for this.

If orientation is right or left then they are returned to normal.

I've bound this to XF86RotateWindows in my XFCE4 keyboard shortcuts, and
I have a xmodmap configuration that sets Hill's rotation button to this
key. Thus I can make Hill rotate her screen with a button press.

The entry from my ~/.Xmodmap file is thus:

	keycode 204 = XF86RotateWindows