Add myworkspace script.
This goes to my workspace directory (or $1) and does a select over
directories found; pick one and it runs mytmux to easily spin up a tmux
session there or connect to one already there.

Because my typical morning looks a lot like this:

	boot
	login
	open terminal
	cd ~/sw/workspace/project
	mytmux
	1
	... other stuff

Adding some setup hooks to mytmux for automatically setting the tabs in
my tmux session might be worth while.