Add a .gitconfig.local.
Looks like includeIf was added back in 2018 as part of Git 2.13.

Makes it an easy way to define a ~/.gitconfig snippet local to the
current machine. Finally, a way to be able to have my .gitconfig
committed and not have to play the
is-it-the-right-e-mail-i-just-committed-with game.

Plus this would be a good way to be able to move the user.name and
user.email out of my tracked .gitconfig, at long last.

For now it's set to sw/workspace, where I typically clone projects.