zsh: emacs line editing.
In Apple Terminal the default behavor seems to be suitably emacs already
that I didn't think changing this was necessary. But in the VS Code
terminal: I end up getting ^P and such inserted instead of the shell's
command line editing.

Setting emacs mode explicitly seems to solve this. So that settles that.