Remove -X from LESS, and document the flags passed.
Removing -X/--no-init because this can prevent using the mouse wheel to
scroll in some newer terminals.

Originally, I had added this so that less wouldn't clear the screen when
exited. But I have to use less without my $LESS often enough on other
machines that I've stopped caring about that and learned to live with it.