vimrc: fix case sensitivity of autocmd filetype.
Looks like this is an error in the old Vim 7.4 on CentOS, or at least
for the 'vim-enhanced' package. Throws an error that the event does not
exist.

Later 8.x and 9.x versions I've got handy don't care about the case.
Making this FileType, since that appeases both and matches the
documentation in autocmd.txt.